+++ CHANGES.md
@@ -3,6 +3,8 @@ E-Maj - Change log
 <devel>
 ------
 ###Enhancements:###
+  * Add an internal table that records log sessions, i.e. time ranges between
+    start_group and stop_group operations.
   * Minor code changes.
 
 ###Bug fixes:###